SPLICE-1500 Extend skipWAL/WWdetection to index writes
