+++ splice_machine/src/main/java/com/splicemachine/derby/stream/output/PermissiveInsertWriteConfiguration.java
@@ -39,14 +39,11 @@ import java.util.concurrent.ExecutionException;
  */
 public class PermissiveInsertWriteConfiguration extends ForwardingWriteConfiguration {
     private static final Logger LOG = Logger.getLogger(PermissiveInsertWriteConfiguration.class);
-    protected boolean skipWAL;
-    protected boolean skipConflictDetection;
     protected OperationContext operationContext;
     protected ThreadLocal<PairDecoder> pairDecoder;
 
     public PermissiveInsertWriteConfiguration(WriteConfiguration delegate, OperationContext operationContext,
-                                              final PairEncoder encoder, final ExecRow execRowDefinition,
-                                              boolean skipConflictDetection, boolean skipWAL) {
+                                              final PairEncoder encoder, final ExecRow execRowDefinition) {
         super(delegate);
         assert operationContext!=null && encoder != null:"Passed in null values to PermissiveInsert";
         this.operationContext = operationContext;
@@ -56,8 +53,6 @@ public class PermissiveInsertWriteConfiguration extends ForwardingWriteConfigura
                 return encoder.getDecoder(execRowDefinition.getClone());
             }
         };
-        this.skipWAL = skipWAL;
-        this.skipConflictDetection = skipConflictDetection;
     }
 
     @Override
@@ -146,16 +141,6 @@ public class PermissiveInsertWriteConfiguration extends ForwardingWriteConfigura
         }
         sb.append(": " + row);
         return sb.toString();
-    }
-
-    @Override
-    public boolean skipConflictDetection() {
-        return skipConflictDetection;
-    }
-
-    @Override
-    public boolean skipWAL() {
-        return skipWAL;
     }
 }
